eighteenth commit with migration scripts modified

+++ src/main/java/com/example/demo_Electricity_App/DTO/Master/Response/CityResponseDTO.java
@@ -7,7 +7,7 @@ public class CityResponseDTO {
     private Long id;
     private String name;
     private boolean isActive;
-    private Integer districtId;
+    private Long districtId;
     private String districtName;
     private Long cityHeadId;
     private String cityHeadName;
